Further test fixes to get toggle functionality to work
Former-commit-id: 08dea8d6e0c8ec660f7ddb31a5fc55b4d07419f4
Former-commit-id: bb311bc595e148161530ba26a5780ceadda8dbef

+++ src/simulation/power/_GeneralModuleFiles/simPowerNodeBase.h
@@ -67,7 +67,7 @@ public:
     int64_t nodePowerOutMsgId;
     int64_t nodeStatusInMsgId;
     double nodePowerOut; //!< [W] Power provided (+) or consumed (-).
-    uint8_t powerStatus; //!< Device power mode; by default, 0 is off and 1 is on. Additional modes can fill other slots
+    uint64_t powerStatus; //!< Device power mode; by default, 0 is off and 1 is on. Additional modes can fill other slots
 
 protected:
     PowerNodeUsageSimMsg nodePowerMsg;
